добавлено множество тестов черех DataRow

+++ AptekaEu/AptekaEuLib/Product.cs
@@ -5,8 +5,8 @@
         public int Id { get; }
         public string Name { get; set; }
         public int CategoryId { get; set; }
-        public decimal PurchasePrice { get; set; }
-        public decimal SalePrice { get; set; }
+        public double PurchasePrice { get; set; }
+        public double SalePrice { get; set; }
         public int ActualQuantity { get; set; }
     }
 }
